Refactor helper scripts to utils module only
Also, add test and plot modules for RL.
