+++ rl_scripts/args/registry_args.py
@@ -1,4 +1,4 @@
-from rl_scripts.helpers.setup_helpers import setup_ppo
+from rl_scripts.utils.setup_helpers import setup_ppo
 from rl_scripts.algorithms.ppo import PPO
 from rl_scripts.helpers.setup_helpers import  setup_A2C
 from rl_scripts.algorithms.a2c import A2C
