+++ rl_scripts/gymnasium_envs/general_sim_env.py
@@ -1,10 +1,10 @@
 import gymnasium as gym
 
 from rl_scripts.utils.env_util import SimEnvUtils
-from rl_scripts.helpers.setup_helpers import setup_rl_sim, SetupHelper
-from rl_scripts.helpers.general_helpers import CoreUtilHelpers, SimEnvHelpers
+from rl_scripts.utils.setup_helpers import setup_rl_sim, SetupHelper
+from rl_scripts.utils.general_helpers import CoreUtilHelpers, SimEnvHelpers
 from rl_scripts.agents.path_agent import PathAgent
-from rl_scripts.helpers.drl_helpers import get_obs_space, get_action_space
+from rl_scripts.utils.drl_helpers import get_obs_space, get_action_space
 
 from rl_scripts.algorithms.algorithm_props import RLProps
