+++ rl_scripts/model_manager.py
@@ -2,7 +2,7 @@ import os
 
 from helper_scripts.sim_helpers import parse_yaml_file
 
-from rl_scripts.helpers.general_helpers import determine_model_type
+from rl_scripts.utils.general_helpers import determine_model_type
 from rl_scripts.args.registry_args import ALGORITHM_REGISTRY
