+++ rl_scripts/utils/deep_rl.py
@@ -1,5 +1,5 @@
 from rl_scripts.args.general_args import VALID_PATH_ALGORITHMS
-from rl_scripts.helpers.general_helpers import determine_model_type
+from rl_scripts.utils.general_helpers import determine_model_type
 from rl_scripts.args.registry_args import ALGORITHM_REGISTRY
